# Optimize differencing in  time