@no-lex 2021-07-16
Performance tests:
Intel Xeon E5 2440 v2 @ 1.9Ghz w/ 6x16GB PC3-10600R
Ubuntu 18.04.5 Python 3.7

**12mb file:**
Old:
 - 92107 function calls (89359 primitive calls) in 0.223 seconds
 
New:
 - 34606 function calls (31750 primitive calls) in 0.103 seconds (2.16x faster)
 
**16gb file:**
Old:
 - 6722358 function calls (6559920 primitive calls) in 93.478 seconds

New:
 - 1004773 function calls (969375 primitive calls) in 46.040 seconds (2.02x faster)

**31gb file:**
Old:
 - 6625565 function calls (6464503 primitive calls) in 197.469 seconds

New:
 - 989541 function calls (954655 primitive calls) in 109.798 seconds (1.79x faster)